When a file is uploaded, wait 60 seconds before fetching stats
We're uploading a YAML file as part of each deploy, and noticing the the
stats we fetch are lower than expected.

Could it be that while the uploaded file is being processed, the number
of phrases in the uploaded locale is under reported? Let's delay
fetching the stats and see what happens